small modifications
