+++ CollectionvViewDemo/Data/HourlyEntry.swift
@@ -1,0 +1,14 @@
+//
+//  HourlyEntry.swift
+//  CollectionvViewDemo
+//
+//  Created by Bence Pattogato on 01/09/2021.
+//
+
+import Foundation
+
+struct HourlyEntry: Hashable {
+    let time: String
+    let icon: WeatherIcon
+    let temperature: String
+}
